Fix trace preflight CI regressions
