+++ src/core/extension/trace/mod.rs
@@ -15,6 +15,7 @@ pub mod baseline;
 mod overlay;
 mod overlay_lock;
 pub mod parsing;
+mod preflight;
 pub mod probes;
 pub mod report;
 pub mod run;
@@ -47,8 +48,7 @@ pub use report::{
 };
 pub use report::{push_overlay_markdown, render_markdown};
 pub use run::{
-    resolve_declared_trace_artifact_path, run_trace_list_workflow, run_trace_workflow,
-    trace_is_unclaimed, TraceListWorkflowArgs,
+    run_trace_list_workflow, run_trace_workflow, trace_is_unclaimed, TraceListWorkflowArgs,
 };
 pub use run::{TraceRunWorkflowArgs, TraceRunWorkflowResult, TraceRunnerInputs};
 pub use span_summary::{attach_span_summary_metadata, TraceSpanSummaryOutput};
